chore: updated readme and appicon, minor refactoring, added default nav view
Updated the readme with some basic data so that it isn't bare and also changed the AppIcon badge to say "ALPHA" rather than the version to avoid having to regenerate a new AppIcon with every build. Refactored LiveParser2 by deleting old non-working code and making it the actual LiveParser. Lastly, changed ContentView so that it defaults to a view based on which ViewState enum value is selected.